+++ carp_core/test/study_json_test.dart
@@ -10,7 +10,7 @@ void main() {
 
   setUp(() {
     study = new Study("1234", "bardram", name: "bardram study");
-    study.dataEndPoint = new FileDataEndPoint(DataEndPointType.FILE);
+    study.dataEndPoint = new FileDataEndPoint();
 
     Task task = new Task("Generic task");
     final Measure m = new Measure(Measure.GENERIC_MEASURE, name: 'Generic measure');