+++ carp_webservices/test/carp_service_test.dart
@@ -65,12 +65,17 @@ void main() {
 
       final CARPDataPoint data = CARPDataPoint.fromDatum(study.id, study.userId, datum);
 
-      //print(_encode(data.toJson()));
+      print(_encode(data.toJson()));
 
       data_point_id = await CarpService.instance.getDataPointReference().postDataPoint(data);
 
       assert(data_point_id.length > 0);
       print("data_point_id : $data_point_id");
+    });
+
+    test('- batch', () async {
+      final File file = File("test/batch.json");
+      await CarpService.instance.getDataPointReference().batchPostDataPoint(file);
     });
 
     test('- get', () async {