+++ carp_backend/CHANGELOG.md
@@ -1,3 +1,2 @@
-## 0.0.1 - Initial version, Oct. 2018.
-
-* Initial version of dedicated package bringing together the old implementation.
+## 0.1.0
+* Initial version.