+++ carp_webservices/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+* Small refactor and removal of debug print.
+* Adding an example
+* Added support for batch upload of data points as a file.
+
 ## 0.1.1
 * Documentation update.